# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      error: {}
      status:
        type: integer
    type: object
...
  models.OrderProductCreate:
    properties:
      order_id:
        type: string
      price:
        type: string
      product_id:
        type: string
      quantity:
        type: string
    type: object
  models.OrderProductUpdate:
    properties:
      id:
        type: string
      order_id:
        type: string
      price:
        type: string
      product_id:
        type: string
      quantity:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /e_commerce/api/v1/order_product:
    get:
      consumes:
      - application/json
      description: Get List OrderProducts
      operationId: get_list_order_products
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: category_id
        in: query
        name: category_id
        type: string
      - description: brand_id
        in: query
        name: brand_id
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List OrderProducts
      tags:
      - OrderProduct
    post:
      consumes:
      - application/json
      description: Create OrderProduct
      operationId: create_order_product
      parameters:
      - description: CreateOrderProductRequest
        in: body
        name: OrderProduct
        required: true
        schema:
          $ref: '#/definitions/models.OrderProductCreate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create OrderProduct
      tags:
      - OrderProduct
  /e_commerce/api/v1/order_product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete OrderProduct
      operationId: delete_order_product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete OrderProduct
      tags:
      - OrderProduct
    get:
      consumes:
      - application/json
      description: Get By ID OrderProduct
      operationId: get_by_id_order_product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID OrderProduct
      tags:
      - OrderProduct
    put:
      consumes:
      - application/json
      description: Update OrderProduct
      operationId: update_order_product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateOrderProductRequest
        in: body
        name: OrderProduct
        required: true
        schema:
          $ref: '#/definitions/models.OrderProductUpdate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update OrderProduct
      tags:
      - OrderProduct
...
swagger: "2.0"